+++ biopandas/__init__.py
@@ -24,5 +24,5 @@
 # 'X.Y.dev0' is the canonical version of 'X.Y.dev'
 #
 
-__version__ = "0.5.1"
+__version__ = "0.6.0dev"
 __author__ = "Sebastian Raschka <mail@sebastianraschka.com>"