+++ biopandas/align/__init__.py
@@ -1,0 +1,15 @@
+# BioPandas
+# Author: Sebastian Raschka <mail@sebastianraschka.com>
+# License: BSD 3 clause
+# Project Website: http://rasbt.github.io/biopandas/
+# Code Repository: https://github.com/rasbt/biopandas
+
+"""
+BioPandas module for working with a collection
+Protein Data Bank (PDB) files.
+"""
+
+from .align import Align
+from .tmalign import TMAlign
+
+__all__ = ['Align', 'TMAlign']