+++ biopandas/stack/__init__.py
@@ -1,0 +1,14 @@
+# BioPandas
+# Author: Sebastian Raschka <mail@sebastianraschka.com>
+# License: BSD 3 clause
+# Project Website: http://rasbt.github.io/biopandas/
+# Code Repository: https://github.com/rasbt/biopandas
+
+"""
+BioPandas module for working with a collection
+Protein Data Bank (PDB) files.
+"""
+
+from .stack import PandasPdbStack
+
+__all__ = ["PandasPdbStack"]