+++ docs/CHANGELOG.md
@@ -5,6 +5,14 @@
 The CHANGELOG for the current development version is available at
 [https://github.com/rasbt/biopandas/blob/main/docs/sources/CHANGELOG.md](https://github.com/rasbt/biopandas/blob/main/docs/sources/CHANGELOG.md).
 
+### 0.6.1dev1 (UNRELEASED)
+- Adds support for constructing a collection of PDB files and perform actions on them (Via [Julia K. Varga](https://github.com/gezmi)
+- Adds the possibility of performing alignments, translation and rotation on the structure, via an extendable Align class (Via [Julia K. Varga](https://github.com/gezmi)
+- TMAlign is installed alongside the package to perform the alignments (Via [Julia K. Varga](https://github.com/gezmi)
+- Added feature to write out mmcif files
+- Fixed erronoues links in documentation, together with automatic generation of said documents (Via [Julia K. Varga](
+- Dev: switched testing framework entirely to pytest. Drops nose dependency due to version conflicts with Python 3.12 (`nose`) and 3.8 (`nose`)
+
 ### 0.5.1 (01/08/2024)
 
 - Fix: improves support for writing PDBs with `OTHERS` records to stream. PR [#149](https://github.com/BioPandas/biopandas/pull/149). Addresses issue [#141](https://github.com/BioPandas/biopandas/issues/141).